Refactor BankAccount methods to use int parameters

+++ BankAccount/IBankAccount.cs
@@ -2,8 +2,8 @@
 {
     public interface IBankAccount
     {
-        void Deposit(double amount);
-        void Withdraw(double amount);
+        void Deposit(int amount);
+        void Withdraw(int amount);
         void PrintStatement();
     }
 }
